// src/app/models/dashobard.model.ts
export class Dashoboard {
    positive: number;
    totalpositive: number;
    recovered: number;
    totalrecovered: number;
    active:number;
    totalactive: number;
    deceased: number;
    totaldeceased: number;
    tested:number;
    totaltested:number;
    modificationdate: Date;
    creationdate: Date;
}